feat:Integration for cal-chat

+++ packages/trpc/server/routers/viewer/eventTypes/delete.handler.ts
@@ -1,4 +1,5 @@
 import { ONEHASH_API_KEY, ONEHASH_CHAT_SYNC_BASE_URL } from "@calcom/lib/constants";
+import { isPrismaObjOrUndefined } from "@calcom/lib/isPrismaObj";
 import { prisma } from "@calcom/prisma";
 
 import type { TrpcSessionUser } from "../../../trpc";
@@ -11,7 +12,7 @@ type DeleteOptions = {
   input: TDeleteInputSchema;
 };
 
-export const deleteHandler = async ({ ctx: _ctx, input }: DeleteOptions) => {
+export const deleteHandler = async ({ ctx, input }: DeleteOptions) => {
   const { id } = input;
 
   await prisma.eventTypeCustomInput.deleteMany({
@@ -26,24 +27,30 @@ export const deleteHandler = async ({ ctx: _ctx, input }: DeleteOptions) => {
     },
   });
 
-  if (_ctx.user.metadata?.oh_chat_enabled) {
-    await handleOHChatSync(id, _ctx.user.email);
+  if (isPrismaObjOrUndefined(ctx.user.metadata)?.connectedChatAccounts) {
+    await handleOHChatSync(id);
   }
   return {
     id,
   };
 };
 
-const handleOHChatSync = async (uid: string, email: string) => {
-  await fetch(`${ONEHASH_CHAT_SYNC_BASE_URL}/cal_event`, {
+const handleOHChatSync = async (eventUid: string) => {
+  const credential = await prisma.credential.findFirst({
+    where: {
+      appId: "onehast-chat",
+      userId: uid,
+    },
+  });
+  const account_user_id: string = isPrismaObjOrUndefined(credential?.key)?.account_user_id.toString();
+  if (!account_user_id) return Promise.resolve(null);
+  const queryParams = new URLSearchParams({ account_user_id: account_user_id, uid: uid });
+
+  await fetch(`${ONEHASH_CHAT_SYNC_BASE_URL}/cal_event?${queryParams}`, {
     method: "DELETE",
     headers: {
       "Content-Type": "application/json",
       Authorization: `Bearer ${ONEHASH_API_KEY}`,
     },
-    body: JSON.stringify({
-      email,
-      uid,
-    }),
   });
 };
